ai_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
numpy
opencv-python-headless
tensorflow
torch
torchvision
scikit-learn
pillow
imagehash
exifread
piexif
python-jose[cryptography]
passlib[bcrypt]
httpx
celery
redis
python-dotenv
pytest
pytest-asyncio
mtcnn
# yolov8 or ultralytics for object detection
ultralytics
